split business component
